Move release prep out of github action into manual script (#30)
For now we need all changes to main to go via PR (security
requirement).

Overall I felt it was simpler to just have the release prepping
be manually triggerd for now (with a script).

I think this is actually nicer in some ways because it means we
can control exactly when we want to cut a release (previously it
would always release as soon as single new changeset hit the main
branch).

I also remove the git tagging - I don't think this really made
sense given this is a monorepo with multiple packages with out of
sync versions.